Provide hint for virtio0 device
This'll suggest use of the `montero-vm-storage-lvm` storage when
importing the VM into proxmox and set the size of the virtual disk to
20GB.

The root filesystem is configured with `autoResize = true`, so it should
take up the full size of the disk on boot before mounting.

If I need to expand the size later, I can just make the LV for the
virtual disk bigger and everything should resize correctly in theory.