Pass stateVerson explicitly
The primary benefit of this change is that it makes it easier to use the
same home-manager configuration for different machines that might have
initially been on different versions of home-manager.

On NixOS machines, it also makes it easier to see that the same
stateVersion is used for both system.stateVersion and for home-manager.