Add oauth2 proxy through Google in front of transmission
Originally, I was using the tailscale authentication proxy, but that
just checks if the request is coming from a Tailscale address. I don't
think that's a strong enough check since basically any program on a
machine in the tailnet would be able to access this service. I'd rather
force authentication through Google, since at least then it means that
the user is likely deliberately accessing the service directly from
their web browser.

I think this breaks using the Tailscale JSON RPC mechanism, but I don't
really use it, so it seems fine for now.