Disable direct access to Plex
This change makes it basically impossible to access Plex without TLS.

Plex can be accessed from inside the tailnet at
https://plex.toothyshouse.com and from outside the tailnet at
https://plex.toothyshouse.com:32400. Inbound NAT translates the former to
hit the LAN ip of the box on port 443, so both options go through the
nginx proxy, which terminates TLS.

The trick was setting "Custom server access URLs" to

    https://plex.toothyshouse.com:443,https://plex.toothyshouse.com:32400

The `:443` is unexpectedly necessary. Otherwise, I think it just
defaults to `:32400`.

I also disabled "Remote Access" from the Plex server settings, since the
check to see if the server is accessible remotely seems to not consider
the "Custom server access URLs". Instead, it just tries to hit the
specified port on the WAN IP of the box. Even then, it gives false
negatives. The setting doesn't seem to actually be important; remote
access actually just works fine with this setting disabled.